+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/view/ListWorksetView.java
@@ -30,6 +30,8 @@ public class ListWorksetView implements Aliasable, Serializable, Timestamped {
   private String alias;
   private int itemCount;
   private String description;
+  private String category;
+  private String stage;
 
   @ManyToOne(targetEntity = UserImpl.class)
   @JoinColumn(name = "creator")
@@ -83,6 +85,22 @@ public class ListWorksetView implements Aliasable, Serializable, Timestamped {
 
   public void setDescription(String description) {
     this.description = description;
+  }
+
+  public String getCategory() {
+    return category;
+  }
+
+  public void setCategory(String category) {
+    this.category = category;
+  }
+
+  public String getStage() {
+    return stage;
+  }
+
+  public void setStage(String stage) {
+    this.stage = stage;
   }
 
   @Override