+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/workset/WorksetCategory.java
@@ -1,0 +1,59 @@
+package uk.ac.bbsrc.tgac.miso.core.data.impl.workset;
+
+import java.io.Serializable;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+import uk.ac.bbsrc.tgac.miso.core.data.Aliasable;
+import uk.ac.bbsrc.tgac.miso.core.data.Deletable;
+
+@Entity
+public class WorksetCategory implements Aliasable, Deletable, Serializable {
+
+  private static final long serialVersionUID = 1L;
+  private static final long UNSAVED_ID = 0L;
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long categoryId;
+
+  private String alias;
+
+  @Override
+  public long getId() {
+    return categoryId;
+  }
+
+  @Override
+  public void setId(long id) {
+    this.categoryId = id;
+  }
+
+  @Override
+  public boolean isSaved() {
+    return getId() != UNSAVED_ID;
+  }
+
+  @Override
+  public String getDeleteType() {
+    return "Workset Category";
+  }
+
+  @Override
+  public String getDeleteDescription() {
+    return getAlias();
+  }
+
+  @Override
+  public String getAlias() {
+    return alias;
+  }
+
+  public void setAlias(String alias) {
+    this.alias = alias;
+  }
+
+}