+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/WorksetCategoryService.java
@@ -1,0 +1,8 @@
+package uk.ac.bbsrc.tgac.miso.core.service;
+
+import uk.ac.bbsrc.tgac.miso.core.data.impl.workset.WorksetCategory;
+
+public interface WorksetCategoryService
+    extends DeleterService<WorksetCategory>, ListService<WorksetCategory>, BulkSaveService<WorksetCategory> {
+
+}