+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/WorksetStageService.java
@@ -1,0 +1,8 @@
+package uk.ac.bbsrc.tgac.miso.core.service;
+
+import uk.ac.bbsrc.tgac.miso.core.data.impl.workset.WorksetStage;
+
+public interface WorksetStageService
+    extends DeleterService<WorksetStage>, ListService<WorksetStage>, BulkSaveService<WorksetStage> {
+
+}