+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/rest/WorksetCategoryRestController.java
@@ -1,0 +1,63 @@
+package uk.ac.bbsrc.tgac.miso.webapp.controller.rest;
+
+import java.io.IOException;
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+import com.fasterxml.jackson.databind.node.ObjectNode;
+
+import uk.ac.bbsrc.tgac.miso.core.data.impl.workset.WorksetCategory;
+import uk.ac.bbsrc.tgac.miso.core.service.WorksetCategoryService;
+import uk.ac.bbsrc.tgac.miso.dto.Dtos;
+import uk.ac.bbsrc.tgac.miso.dto.SimpleAliasableDto;
+import uk.ac.bbsrc.tgac.miso.webapp.controller.ConstantsController;
+import uk.ac.bbsrc.tgac.miso.webapp.controller.component.AsyncOperationManager;
+
+@Controller
+@RequestMapping("/rest/worksetcategories")
+public class WorksetCategoryRestController extends RestController {
+
+  @Autowired
+  private WorksetCategoryService worksetCategoryService;
+  @Autowired
+  private AsyncOperationManager asyncOperationManager;
+  @Autowired
+  private ConstantsController constantsController;
+
+  @PostMapping(value = "/bulk-delete")
+  @ResponseBody
+  @ResponseStatus(HttpStatus.NO_CONTENT)
+  public void bulkDelete(@RequestBody(required = true) List<Long> ids) throws IOException {
+    RestUtils.bulkDelete("Workset Category", ids, worksetCategoryService);
+    constantsController.refreshConstants();
+  }
+
+  @PostMapping("/bulk")
+  @ResponseStatus(HttpStatus.ACCEPTED)
+  public @ResponseBody ObjectNode bulkCreateAsync(@RequestBody List<SimpleAliasableDto> dtos) throws IOException {
+    return asyncOperationManager.startAsyncBulkCreate("Workset Category", dtos, Dtos::toWorksetCategory, worksetCategoryService, true);
+  }
+
+  @PutMapping("/bulk")
+  @ResponseStatus(HttpStatus.ACCEPTED)
+  public @ResponseBody ObjectNode bulkUpdateAsync(@RequestBody List<SimpleAliasableDto> dtos) throws IOException {
+    return asyncOperationManager.startAsyncBulkUpdate("Workset Category", dtos, Dtos::toWorksetCategory, worksetCategoryService, true);
+  }
+
+  @GetMapping("/bulk/{uuid}")
+  public @ResponseBody ObjectNode getProgress(@PathVariable String uuid) throws Exception {
+    return asyncOperationManager.getAsyncProgress(uuid, WorksetCategory.class, worksetCategoryService, Dtos::asDto);
+  }
+
+}