+++ sqlstore/src/it/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateListWorksetViewDaoIT.java
@@ -1,8 +1,7 @@
 package uk.ac.bbsrc.tgac.miso.persistence.impl;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
-import java.io.IOException;
 import java.util.List;
 
 import org.hibernate.SessionFactory;
@@ -12,7 +11,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import uk.ac.bbsrc.tgac.miso.AbstractDAOTest;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.view.ListWorksetView;
-import uk.ac.bbsrc.tgac.miso.core.util.PaginationFilter;
 
 public class HibernateListWorksetViewDaoIT extends AbstractDAOTest {
 
@@ -31,29 +29,6 @@ public class HibernateListWorksetViewDaoIT extends AbstractDAOTest {
   public void testListBySearch() throws Exception {
     List<ListWorksetView> results = sut.listBySearch("two");
     assertEquals(1, results.size());
-  }
-
-  @Test
-  public void testSearchByCreator() throws IOException {
-    testSearch(PaginationFilter.user("admin", true));
-  }
-
-  @Test
-  public void testSearchByModifier() throws IOException {
-    testSearch(PaginationFilter.user("admin", false));
-  }
-
-  /**
-   * Verifies Hibernate mappings by ensuring that no exception is thrown by a search
-   * 
-   * @param filter the search filter
-   * @throws IOException
-   */
-  private void testSearch(PaginationFilter filter) throws IOException {
-    // verify Hibernate mappings by ensuring that no exception is thrown
-    assertNotNull(sut.list(err -> {
-      throw new RuntimeException(err);
-    }, 0, 10, true, "id", filter));
   }
 
 }