+++ sqlstore/src/it/resources/db/migration/V9000__clear_data.sql
@@ -118,10 +118,12 @@ DELETE FROM Transfer_Sample;
 DELETE FROM User;
 DELETE FROM User_Group;
 DELETE FROM Workset;
+DELETE FROM WorksetCategory;
 DELETE FROM WorksetChangeLog;
 DELETE FROM Workset_Library;
 DELETE FROM Workset_LibraryAliquot;
 DELETE FROM Workset_Sample;
+DELETE FROM WorksetStage;
 DELETE FROM Workstation;
 
 SET FOREIGN_KEY_CHECKS=1;