+++ sqlstore/src/main/resources/db/migration_afterMigrate/05_view_workset.sql
@@ -4,11 +4,15 @@ SELECT
   w.alias,
   (COALESCE(sam.count, 0) + COALESCE(lib.count, 0) + COALESCE(ali.count, 0)) AS itemCount,
   w.description,
+  cat.alias AS category,
+  stage.alias AS stage,
   w.creator,
   w.created,
   w.lastModifier,
   w.lastModified
 FROM Workset w
+LEFT JOIN WorksetCategory cat ON cat.categoryId = w.categoryId
+LEFT JOIN WorksetStage stage ON stage.stageId = w.stageId
 LEFT JOIN (
   SELECT worksetId, COUNT(sampleId) AS count
   FROM Workset_Sample