+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 Changes:
 
+* Added workset categories and stages
 * Added detailed QC statuses for run-libraries
 * Prevent Run Scanner data from overriding user-specified index sequencing or
   sequencing parameters